chore(tests): move view rendering tests to unit suite
Moves view rendering tests to static analysis, as they do not require
a database connection
