+++ engine/tests/phpunit/unit/Elgg/Views/ListingViewsRenderingTest.php
@@ -2,8 +2,6 @@
 
 namespace Elgg\Views;
 
-use Elgg\ViewRenderingTestCase;
-
 /**
  * @group ViewRendering
  * @group ViewsService
